+++ apps/theme/components/Color/Color.tsx
@@ -1,5 +1,5 @@
 import { omit } from '@digdir/designsystemet-react';
-import type { ColorInfo, ColorType } from '@digdir/designsystemet/color';
+import type { ColorInfo, ThemeColors } from '@digdir/designsystemet/color';
 import { SunIcon } from '@navikt/aksel-icons';
 import cl from 'clsx/lite';
 import { forwardRef } from 'react';
@@ -16,7 +16,7 @@ type ColorProps = {
   featured?: boolean;
   hex?: string;
   showColorMeta?: boolean;
-  type: ColorType;
+  type: ThemeColors;
 } & Omit<React.HTMLAttributes<HTMLButtonElement>, 'color'>;
 
 const Color = forwardRef<HTMLButtonElement, ColorProps>(