+++ packages/react/src/components/Table/Table.stories.tsx
@@ -54,6 +54,7 @@ Preview.args = {
   zebra: false,
   stickyHeader: false,
   border: false,
+  hover: false,
 };
 
 const dummyData = [
@@ -256,6 +257,35 @@ export const WithFormElements: Story = (args) => {
             <TableCell>
               <Textfield size='sm' />
             </TableCell>
+          </TableRow>
+        ))}
+      </TableBody>
+    </Table>
+  );
+};
+
+export const FixedTable: Story = (args) => {
+  const rows = Array.from({ length: 3 }, (_, i) => i + 1);
+  return (
+    <Table
+      {...args}
+      style={{
+        tableLayout: 'fixed',
+      }}
+    >
+      <TableHead>
+        <TableRow>
+          <TableHeaderCell>Header 1</TableHeaderCell>
+          <TableHeaderCell>Header 2</TableHeaderCell>
+          <TableHeaderCell>Header 3</TableHeaderCell>
+        </TableRow>
+      </TableHead>
+      <TableBody>
+        {rows.map((row) => (
+          <TableRow key={row}>
+            <TableCell>{`Cell ${row}1`}</TableCell>
+            <TableCell>{`Cell ${row}2`}</TableCell>
+            <TableCell>{`Cell ${row}3`}</TableCell>
           </TableRow>
         ))}
       </TableBody>