+++ packages/react/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './Button';
+export * from './Badge';
 export * from './Breadcrumbs';
 export * from './HelpText';
 export * from './Spinner';